chore: move enums' location
