--- src/main/java/com/example/comitserver/entity/Position.java
@@ -1,6 +1,0 @@
-package com.example.comitserver.entity;
-
-public enum Position {
-    LEADER,
-    MEMBER
-}
