+++ src/main/java/com/example/comitserver/entity/Progress.java
@@ -1,5 +1,6 @@
 package com.example.comitserver.entity;
 
+import com.example.comitserver.entity.enumeration.Status;
 import jakarta.persistence.*;
 import lombok.*;
