+++ src/main/java/com/example/comitserver/entity/enumeration/Position.java
@@ -1,0 +1,6 @@
+package com.example.comitserver.entity.enumeration;
+
+public enum Position {
+    LEADER,
+    MEMBER
+}
