+++ src/main/java/com/example/comitserver/entity/enumeration/Status.java
@@ -1,4 +1,4 @@
-package com.example.comitserver.entity;
+package com.example.comitserver.entity.enumeration;
 
 public enum Status {
     NOT_STARTED,
